refactor(install): split the harness table from its file primitives
`integrations.rs` was doing three jobs: the harness table, a
read-modify-write pair per config format, and the record of which files
install created. At 591 production lines and a cyclomatic total of 140 it
was the module's worst file by a factor of four.

Same split `render.rs` documents making out of `evaluate.rs` — "printing
is a separate concern from `run`'s ... orchestration, and bundling both
pushed the file's cyclomatic total over the SIMPLE gate":

  integrations.rs  the HARNESSES table + Artifact and its dispatch
  edits.rs         one read-modify-write pair per config format
  ownership.rs     which files install created; owned deletion

Pure code movement — no behavior change, the 74 tests are unchanged apart
from moving next to what they cover. `delete_if_empty_and_owned` and
`delete_text_if_blank_and_owned` now share their guard through one
`delete_if_owned`, since they differed only in the emptiness predicate.
Test scratch dirs move to `testing.rs` and get an atomic counter instead
of a label-length suffix, so two tests sharing a label cannot collide.

Production code with tests stripped from both sides — worst file 140 → 90,
SIMPLE average 21% → 27%, and no single function above 7 outside the
untouched `menu.rs`. Full numbers in docs/decisions/cli-harness-install.md.
